+++ Examples/LatchwayChat/index.js
@@ -2,7 +2,7 @@
  * @format
  */
 
-import '@latchway/react-native/polyfills';
+import './src/runtime/polyfills';
 import { AppRegistry } from 'react-native';
 import App from './App';
 import { name as appName } from './app.json';